+++ src/migrations/m180412_092824_user_security_columns.php
@@ -12,8 +12,8 @@ class m180412_092824_user_security_columns extends Migration
      */
     public function safeUp()
     {
-        $this->addColumn('admin_user', 'email_verification_token', $this->string(12));
-        $this->addColumn('admin_user', 'email_verification_token_expiration', $this->integer());
+        $this->addColumn('admin_user', 'email_verification_token', $this->string(40));
+        $this->addColumn('admin_user', 'email_verification_token_timestamp', $this->integer());
         $this->addColumn('admin_user', 'login_attempt', $this->integer());
         $this->addColumn('admin_user', 'login_attempt_lock_expiration', $this->integer());
     }
@@ -24,7 +24,7 @@ class m180412_092824_user_security_columns extends Migration
     public function safeDown()
     {
         $this->dropColumn('admin_user', 'email_verification_token');
-        $this->dropColumn('admin_user', 'email_verification_token_expiration');
+        $this->dropColumn('admin_user', 'email_verification_token_timestamp');
         $this->dropColumn('admin_user', 'login_attempt');
         $this->dropColumn('admin_user', 'login_attempt_lock_expiration');
     }